+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
+- (**Migration**) Fix bulk search stopping before all sources were searched in case a match got found but did not fulfill the search criteria (caused entries to require manual review even though a fitting match existed in a lower priority source)
 - (**Browse**) Fix filtering with nested group changes
 - (**Library**) Fix source filter potentially causing all manga to get filtered out
 - (**Chapter**) Fix missing chapters before the first available one not being considered in the total missing chapter count (Example: chapters 1-5 are missing, first available chapter is 6. Chapters 1-5 weren't counted as missing in the total missing chapter amount)